Fixing scaling and rotation of symbols and images when printing the sky image.
There were missing "scale" paameters in some draw functions, which was okay
for screen draws because the default value of 1.0 is correct in that case.
Not so for printed images.

Also added a separate fix to rescale the star pixmaps properly on the printed
image.

Fixes to SkyMap::findPA() to correctly determine an object's position angle

Made SkyMap::pa() virtual so that the overridden versions in KSPlanetBase and
DeepSkyObject will be picked up correctly.

With these fixes, the Print function works again (these problems were all
regressions)

Oh, and in case you're wondering how slow it is to bugfix through an SSH
tunnel that spans continents, the answer is: very, very slow.  :)
I was able to work on this because the wind got too strong last night, so we
had to close the telescope.

CCMAIL: kstars-devel@kde.org

svn path=/trunk/kdeedu/kstars/; revision=334906

+++ kstars/deepskyobject.cpp
@@ -42,7 +42,7 @@ DeepSkyObject::DeepSkyObject( DeepSkyObject &o )
 
 DeepSkyObject::DeepSkyObject( int t, dms r, dms d, float m,
 			QString n, QString n2, QString lname, QString cat,
-			float a, float b, short int pa, int pgc, int ugc )
+			float a, float b, double pa, int pgc, int ugc )
 	: SkyObject( t, r, d, m, n, n2, lname ) {
 	MajorAxis = a;
 	MinorAxis = b;
@@ -55,7 +55,7 @@ DeepSkyObject::DeepSkyObject( int t, dms r, dms d, float m,
 
 DeepSkyObject::DeepSkyObject( int t, double r, double d, float m,
 			QString n, QString n2, QString lname, QString cat,
-			float a, float b, short int pa, int pgc, int ugc )
+			float a, float b, double pa, int pgc, int ugc )
 	: SkyObject( t, r, d, m, n, n2, lname ) {
 	MajorAxis = a;
 	MinorAxis = b;
